Export Stripe variables
Export stripe variables and successfully checkout without handling webhook.

+++ src/Entity/Product.php
@@ -41,7 +41,7 @@ class Product
     private $systemCode;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $price;
 
@@ -80,12 +80,12 @@ class Product
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(int $price): self
+    public function setPrice(float $price): self
     {
         $this->price = $price;
